Massively performance of gathering coverage thanks to improved multi threading.
Previously both the GCC and LLVM generators simply partitioned the amount of files and functions into equally sized batches and then processed them. This lead to harsh thread under utilization however as not every file or function is equally sized. Instead, we now use a ConcurrentLinkedQueue which is a lockless FIFO datastructures which contains all functions that should be processed. Every thread simply pops off functions from the front of the queue for processing until the queue is empty. This leads to practically 100% thread utilization